+++ CHANGELOG.md
@@ -1,4 +1,20 @@
 # 更新日志\n
+## [ForwardWidgets-huangxd] - 2025-07-12 03:18:40
+- 仓库: https://github.com/huangxd-/ForwardWidgets
+- 分支: main
+- 新增: 113710e241ac7cd2fe4375a23162834299c56862
+
+## [ForwardWidgets-2kuai] - 2025-07-12 03:18:40
+- 仓库: https://github.com/2kuai/ForwardWidgets
+- 分支: main
+- 新增: 2b990b15024cc718f476ddcfd36efb3719257ed9
+
+## [ForwardWidgets-pack1r] - 2025-07-12 03:18:40
+- 仓库: https://github.com/pack1r/ForwardWidgets
+- 分支: main
+- 新增: 64224559c8760133446b8beeb338503413174ade
+
+
 ## [ForwardWidgets-huangxd] - 2025-07-11 03:20:35
 - 仓库: https://github.com/huangxd-/ForwardWidgets
 - 分支: main